feat(invoices): show buttons fix
