+++ packages/vendure-plugin-invoices/test/dev-server.ts
@@ -64,11 +64,11 @@ require('dotenv').config();
       AdminUiPlugin.init({
         port: 3002,
         route: 'admin',
-        // app: compileUiExtensions({
-        //   outputPath: path.join(__dirname, '__admin-ui'),
-        //   extensions: [InvoicePlugin.ui],
-        //   devMode: true,
-        // }),
+        app: compileUiExtensions({
+          outputPath: path.join(__dirname, '__admin-ui'),
+          extensions: [InvoicePlugin.ui],
+          devMode: true,
+        }),
       }),
     ],
     paymentOptions: {
